adding APIClient to use for both span/metric export
